confirm added to delete note button

+++ app/public/script.js
@@ -109,15 +109,18 @@ recognition.onresult = function(event) { //the event holds the results
                 var deleteButton = $("<button/>", {
                     text: "Remove Note",
                     click: function() {
-                        $(this).parent().remove();
-                        $.ajax({
-                            url: '/notes/' + data._id,
-                            type: 'DELETE',
-                            success: function(result) {
-                                // Do something with the result
-                                console.log(result);
-                            }
-                        });
+                        if(confirm("Are you sure you want to delete this task?")) {
+                            $(this).parent().remove();
+                            $.ajax({
+                                url: '/notes/' + data._id,
+                                type: 'DELETE',
+                                success: function(result) {
+                                    // Do something with the result
+                                    console.log(result);
+                                }
+                            });    
+                        }
+                        
                         
                     }
                 });
